app/requirements.txt
gunicorn
gevent
tensorflow
keras
Flask
Pillow
numpy
opencv-python
pymongo
dnspython